update the messsage on LHN
